Fix PinnedImages tests for SNO/compact clusters
This commit addresses two issues raised in PR review:

1. Exclude nodes with master/control-plane role from custom MCPs
   - Created GetReadySchedulableWorkerNodes() in test/extended/util package
   - Filters out nodes that have both worker and master/control-plane labels
   - In SNO and compact clusters, nodes belong to master MCP and cannot be moved
     to custom MCPs
   - Prevents test failures when trying to assign these nodes to custom pools

2. Skip test instead of error when insufficient worker nodes
   - Changed addWorkerNodesToCustomPool() to call g.Skip() instead of returning error
   - Insufficient worker nodes is an environmental constraint, not a test failure
   - Provides clearer test results for SNO/compact cluster scenarios

3. Fix import cycle
   - Moved GetReadySchedulableWorkerNodes to util package to avoid import cycle
   - machine_config package can now safely import from util without creating cycles

Tested on compact cluster (3 nodes with worker+master+control-plane roles):
- GetReadySchedulableWorkerNodes() correctly returned 0 nodes
- Tests properly skipped with message: "Not enough schedulable worker nodes (0) for custom MCP (need 1)"
